Use usize indexes
This allows for unioning bigger sets, but also doesn't change the
storage size because it may still downsize in from_index

+++ src/unify/mod.rs
@@ -63,9 +63,9 @@ mod tests;
 pub trait UnifyKey: Copy + Clone + Debug + PartialEq {
     type Value: UnifyValue;
 
-    fn index(&self) -> u32;
+    fn index(&self) -> usize;
 
-    fn from_index(u: u32) -> Self;
+    fn from_index(u: usize) -> Self;
 
     fn tag() -> &'static str;
 
@@ -302,7 +302,7 @@ impl<S: UnificationStore> UnificationTable<S> {
     /// Returns the keys of all variables created since the `snapshot`.
     pub fn vars_since_snapshot(&self, snapshot: &Snapshot<S>) -> Range<S::Key> {
         let range = self.values.values_since_snapshot(&snapshot.snapshot);
-        S::Key::from_index(range.start as u32)..S::Key::from_index(range.end as u32)
+        S::Key::from_index(range.start)..S::Key::from_index(range.end)
     }
 }
 
@@ -318,7 +318,7 @@ impl<S: UnificationStoreMut> UnificationTable<S> {
     /// Creates a fresh key with the given value.
     pub fn new_key(&mut self, value: S::Value) -> S::Key {
         let len = self.values.len();
-        let key: S::Key = UnifyKey::from_index(len as u32);
+        let key: S::Key = UnifyKey::from_index(len);
         self.values.push(VarValue::new_var(key, value));
         debug!("{}: created new key: {:?}", S::tag(), key);
         key
@@ -335,7 +335,7 @@ impl<S: UnificationStoreMut> UnificationTable<S> {
     /// the closure.
     pub fn reset_unifications(&mut self, mut value: impl FnMut(S::Key) -> S::Value) {
         self.values.reset_unifications(|i| {
-            let key = UnifyKey::from_index(i as u32);
+            let key = UnifyKey::from_index(i);
             let value = value(key);
             VarValue::new_var(key, value)
         });
