+++ src/unify/tests.rs
@@ -23,14 +23,14 @@ use unify::{EqUnifyValue, InPlace, InPlaceUnificationTable, NoError, UnifyKey, U
 use unify::{UnificationStore, UnificationTable};
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-struct UnitKey(u32);
+struct UnitKey(usize);
 
 impl UnifyKey for UnitKey {
     type Value = ();
-    fn index(&self) -> u32 {
+    fn index(&self) -> usize {
         self.0
     }
-    fn from_index(u: u32) -> UnitKey {
+    fn from_index(u: usize) -> UnitKey {
         UnitKey(u)
     }
     fn tag() -> &'static str {
@@ -242,14 +242,14 @@ fn even_odd() {
 }
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-struct IntKey(u32);
+struct IntKey(usize);
 
 impl UnifyKey for IntKey {
     type Value = Option<i32>;
-    fn index(&self) -> u32 {
+    fn index(&self) -> usize {
         self.0
     }
-    fn from_index(u: u32) -> IntKey {
+    fn from_index(u: usize) -> IntKey {
         IntKey(u)
     }
     fn tag() -> &'static str {
@@ -363,17 +363,17 @@ fn unify_root_value_2() {
 }
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
-struct OrderedKey(u32);
+struct OrderedKey(usize);
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-struct OrderedRank(u32);
+struct OrderedRank(usize);
 
 impl UnifyKey for OrderedKey {
     type Value = OrderedRank;
-    fn index(&self) -> u32 {
+    fn index(&self) -> usize {
         self.0
     }
-    fn from_index(u: u32) -> OrderedKey {
+    fn from_index(u: usize) -> OrderedKey {
         OrderedKey(u)
     }
     fn tag() -> &'static str {
